refactor(tailwind-variants): modularize core logic and refactor `tv` function for maintainability
Extracted core logic into `tv.ts`, modularizing the implementation of `tv` and related functions to separate concerns and enhance code clarity. Refactored type handling, configuration merging, and slot-specific logic to simplify maintenance and improve type safety across the package.
